requirements.txt
# --- Core Web Framework ---
Django>=5.0
djangorestframework
djangorestframework-simplejwt
python-dotenv
django-crispy-forms
crispy-bootstrap5
django-cors-headers

# --- Data Science & ML ---
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
joblib>=1.3.0
shap>=0.44.0

# --- Visualization & Frontend Helpers ---
plotly>=5.18.0
streamlit>=1.30.0
requests

# --- Deployment ---
gunicorn
whitenoise
dj-database-url
psycopg2-binary
